oa_backend/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-auth-ldap==1.2.16
django-celery==3.2.2
django-ldapdb==0.9.0
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
kombu==3.0.37
olefile==0.44
Pillow==4.3.0
pycrypto==2.6.1
pyldap==2.4.37
PyMySQL==0.7.11
pytz==2017.3
raven==6.3.0
redis==2.10.6
requests==2.18.4
urllib3==1.22
xlwt==1.3.0
xlrd==1.1.0
boto3==1.4.8

